requirements.txt
appdirs==1.4.4
argcomplete==3.5.3
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.1.6
django-cors-headers==4.7.0
django-js-asset==3.0.1
django-mptt==0.16.0
djangorestframework==3.15.2
gunicorn==23.0.0
idna==3.10
importlib_resources==6.5.2
packaging==24.2
pillow==11.1.0
psycopg2-binary==2.9.10
pyngrok==7.2.3
python-barcode==0.15.1
python-decouple==3.8
python-escpos==3.1
PyYAML==6.0.2
qrcode==8.0
requests==2.32.3
setuptools==75.8.0
six==1.17.0
sqlparse==0.5.3
urllib3==2.3.0
